websocket/requirements.txt
pika==1.3.1
flask-sock==0.6.0
gunicorn==20.1.0
pydantic==1.10.4
python-dotenv==0.19.0
pyjwt[crypto]==2.5.0
json-logging==1.3.0
gevent==22.10.2
greenlet==2.0.2